Moved creation of div elements into executeWheel(), using d3
Added a parameter to the function definition to enable mouse-over
functionality.

Also made sure not trying to dereference 'content' if it hasn't been
defined, i.e., if there is no mouse-over annotation box in use.

Since I'm referencing the default hover text in multiple places now,
defined it as a variable. Default text is now always displayed if there
is no title text to be displayed.